(svn r23104) -Codechange: prepare the vehicle/sign z for some further changes to reduce casting

+++ src/signs_base.h
@@ -25,7 +25,7 @@ struct Sign : SignPool::PoolItem<&_sign_pool> {
 	ViewportSign sign;
 	int32        x;
 	int32        y;
-	byte         z;
+	int32        z;
 	OwnerByte    owner; // placed by this company. Anyone can delete them though. OWNER_NONE for gray signs from old games.
 
 	Sign(Owner owner = INVALID_OWNER);
